% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.esm.R
\name{eq.esm}
\alias{eq.esm}
\alias{print.eq.esm}
\title{Test for Evaluating Substantial Mediation}
\usage{
eq.esm(
  x,
  y,
  m,
  alpha = 0.05,
  minc = 0.15,
  eil = -0.15,
  eiu = 0.15,
  nboot = 50L,
  data = NULL,
  ...
)

\method{print}{eq.esm}(x, ...)
}
\arguments{
\item{x}{object of class \code{eq.esm}}

\item{y}{outcome variable}

\item{m}{mediator variable}

\item{alpha}{alpha level (default = .05)}

\item{minc}{minimum correlation between x and y (default is .15)}

\item{eil}{lower bound of equivalence interval(default is -.15)}

\item{eiu}{upper bound of equivalence interval (default is .15)}

\item{nboot}{number of boostraps (default = 50L)}

\item{data}{optional data argument}

\item{...}{extra arguments}
}
\value{
returns a \code{list} containing each analysis and their respective statistics
  and decision


}
\description{
Function computes the equivalence testing method (total effect) for evaluating substantial mediation and Kenny method for full mediation.
}
\examples{
\dontrun{
#equivalence test for substantial mediation
with an equivalenceinterval of -.15 to .15
x<-rnorm(200,sd=2)
m<-.5*x + rnorm(100)
y<-.5*m + rnorm(100)
eq.esm(x,y,m, eil = -.15, eiu = .15)
}
}
\author{
Rob Cribbie \email{cribbie@yorku.ca} and
  Nataly Beribisky \email{natalyb1@yorku.ca}
}
