% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.indvars.R
\name{eq.indvars}
\alias{eq.indvars}
\alias{print.eq.indvars}
\title{Title Function for Evaluating the Equivalence of the Variances of Independent Populations}
\usage{
eq.indvars(dv, iv, eps = 0.5, alpha = 0.05, na.rm = TRUE, data = NULL, ...)

\method{print}{eq.indvars}(x, ...)
}
\arguments{
\item{dv}{Outcome Variable}

\item{iv}{Independent Variable}

\item{eps}{Used to Establish the Equivalence Interval}

\item{alpha}{Nominal Type I Error Rate}

\item{na.rm}{Missing Data Treatment}

\item{data}{Dataset containing dv and iv}

\item{...}{Extra arguments}

\item{x}{object of class \code{eq.indvars}}
}
\value{
returns a \code{list} containing each analysis and their respective statistics
  and decision


}
\description{
Title Function for Evaluating the Equivalence of the Variances of Independent Populations
}
\examples{
\dontrun{
#Two Group Example
indvar<-rep(c("a","b"),c(10,12))
depvar<-rnorm(22)
d<-data.frame(indvar,depvar)
var_equiv(depvar,indvar)
var_equiv(dv=depvar,iv=indvar,eps=.25,data=d)
var_equiv(dv=depvar,iv=indvar,eps=.5)

#Four Group Example
indvar<-rep(c("a","b","c","d"),c(10,12,15,13))
depvar<-rnorm(50)
d<-data.frame(indvar,depvar)
var_equiv(dv=depvar,iv=indvar,eps=.25,data=d)
var_equiv(dv=depvar,iv=indvar)
}
}
\author{
Rob Cribbie \email{cribbie@yorku.ca} and
  Constance Mara \email{Constance.Mara@cchmc.org}
}
