% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfectionism.R
\docType{data}
\name{perfectionism}
\alias{perfectionism}
\title{Perfectionism Data}
\format{
A data frame with 83 rows and 17 variables:
\describe{
  \item{group}{whether the participants received the CBT therapy, a general stress reduction protocol, or no treatment}
  \item{mpshfpre.sop}{Pretest Score, Self-oriented Perfectionism, Hewitt & Flett Multidimensional Perfectionism Scale}
  \item{mpshfpre.spp}{Pretest Score, Socially-prescribed Perfectionism, Hewitt & Flett Multidimensional Perfectionism Scale}
  \item{pcipre.total}{Pretest Score, Perfection Cognitions Inventory}
  \item{baipre.total}{Pretest Score, Beck Anxiety Inventory}
  \item{cesdpre.total}{Pretest Score, CESD Depression Scale}
  \item{mpsfpre.cm}{Pretest Score, Concern Over Mistakes subscale, Frost Multidimensional Perfectionism Scale}
  \item{mpshfpost.sop}{Posttest Score, Self-oriented Perfectionism, Hewitt & Flett Multidimensional Perfectionism Scale}
  \item{mpshfpost.spp}{Posttest Score, Self-prescribed Perfectionism, Hewitt & Flett Multidimensional Perfectionism Scale}
  \item{pcipost.total}{Posttest Score, Perfection Cognitions Inventory}
  \item{baipost.total}{Posttest Score, Beck Anxiety Inventory}
  \item{cesdpost.total}{Posttest Score, CESD Depression Scale}
  \item{mpsfpost.cm}{Posttest Score, Concern Over Mistakes subscale, Frost Multidimensional Perfectionism Scale}
  \item{atqpre.total}{Pretest Score, Automatic Thoughts Quesionnaire}
  \item{atqpost.total}{Posttest Score, Automatic Thoughts Questionnaire}
  \item{mpshfpre.oop}{Pretest score, Other Oriented Perfectionism, Hewitt & Flett Multidimensional Perfectionism Scale}
  \item{mpshfpost.oop}{Posttest Score, Other Oriented Perfectionism, Hewitt & Flett Multidimensional Perfectionism Scale}
  ...
}
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/22122217/}
}
\usage{
perfectionism
}
\description{
This dataset comes from the dissertation of Chantal Arpin-Cribbie.
The study was an RCT looking at the effect of an online CBT therapy on
perfectionism (and related variables) in a sample of undergraduate
students with extreme perfectionism. This dataset has missing data
imputed with a single stochastic regression imputation.
}
\keyword{datasets}
