% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.eq.cat.R
\name{print.eq.cat}
\alias{print.eq.cat}
\title{Print function for the eq.cat function}
\usage{
\method{print}{eq.cat}(x, ...)
}
\arguments{
\item{x}{Data frame from eq.cat}

\item{...}{extra arguments}
}
\value{

}
\description{
Print function for the eq.cat function
}
\examples{
\dontrun{
}
}
