% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.neg.cat.R
\name{print.neg.cat}
\alias{print.neg.cat}
\title{Print function for the neg.cat function}
\usage{
\method{print}{neg.cat}(x, ...)
}
\arguments{
\item{x}{Data frame from neg.cat}

\item{...}{extra arguments}
}
\value{

}
\description{
Print function for the neg.cat function
}
\examples{
\dontrun{
}
}
