% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neg.cat.R
\name{neg.cat}
\alias{neg.cat}
\title{Equivalence Testing for Categorical Variables}
\usage{
neg.cat(x = NULL, y = NULL, tab = NULL, eiU = 0.2, data = NULL, alpha = 0.05)
}
\arguments{
\item{x}{first categorical variable}

\item{y}{second categorical variable}

\item{tab}{contingency table for the two predictor variables}

\item{eiU}{upper limit of equivalence interval}

\item{data}{data file containing the categorical variables}

\item{alpha}{nominal acceptable Type I error rate level}
}
\value{

}
\description{
Testing for the presence of a negligible association between two categorical variables
}
\examples{
#Example 1
x<-rbinom(10,1,.5)
y<-rbinom(10,1,.5)
neg.cat(x,y)
#Example 2
sex<-rep(c("m","f"),c(12,22))
haircol<-rep(c("bld","brn","bld","brn"),c(9,7,11,7))
d <- data.frame(sex,haircol)
tab<-table(sex,haircol)
neg.cat(tab=tab, alpha=.05)
neg.cat(x=sex,y=haircol)
neg.cat(x=sex,y=haircol,data=d)
}
