% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neg.pd.R
\name{neg.pd}
\alias{neg.pd}
\title{Proportional Distance Function (post hoc function - not to be used independently)}
\usage{
neg.pd(effect, PD, EIsign, PDcil, PDciu, cil, ciu, Elevel, Plevel, save)
}
\arguments{
\item{effect}{observed effect}

\item{PD}{proportional distance for effect}

\item{EIsign}{equivalence interval value of the same sign as the effect}

\item{PDcil}{lower bound of the CI for the proportional distance}

\item{PDciu}{upper bound of the CI for the proportional distance}

\item{cil}{lower bound of the CI for the effect}

\item{ciu}{upper bound of the CI for the effect}

\item{Elevel}{1-2alpha CI for the effect}

\item{Plevel}{1-alpha CI for the PD}

\item{save}{Whether to save the plot or not}
}
\value{
nothing is returned
}
\description{
Proportional Distance Function (post hoc function - not to be used independently)
}
\examples{
\dontrun{
1+1
}
}
